Changed drawText() / Added several text / Changed key binding
In the class Text (Text.h/Text.cpp), function drawText() will now take in the font size and color as a parameter.
Added several texts such as "Search Time", "Selection", "Start Point", & "Destination" to the screen. The algorithm is also now timed in seconds.
Added a notice message to the middle of the map when trying to set a start point or destination while wall creation is enabled.